+++ generators/csharp/sdk/src/endpoint/request/WrappedEndpointRequest.ts
@@ -2,6 +2,7 @@ import { csharp } from "@fern-api/csharp-codegen";
 import {
     HttpEndpoint,
     HttpHeader,
+    HttpService,
     Name,
     QueryParameter,
     SdkRequest,
@@ -107,12 +108,14 @@ export class WrappedEndpointRequest extends EndpointRequest {
     }
 
     public getHeaderParameterCodeBlock(): HeaderParameterCodeBlock | undefined {
-        if (this.endpoint.headers.length === 0) {
+        const service = this.getService();
+        const headers = [...this.endpoint.headers, ...service.headers];
+        if (headers.length === 0) {
             return undefined;
         }
         const requiredHeaders: HttpHeader[] = [];
         const optionalHeaders: HttpHeader[] = [];
-        for (const header of this.endpoint.headers) {
+        for (const header of headers) {
             if (this.context.isOptional(header.valueType)) {
                 optionalHeaders.push(header);
             } else {
@@ -208,6 +211,14 @@ export class WrappedEndpointRequest extends EndpointRequest {
         } else {
             return csharp.codeblock(`${parameter}.ToString()`);
         }
+    }
+
+    private getService(): HttpService {
+        const service = this.context.ir.services[this.serviceId];
+        if (service == null) {
+            throw new Error("Unexpected service not found");
+        }
+        return service;
     }
 
     public getRequestBodyCodeBlock(): RequestBodyCodeBlock | undefined {