+++ seed/csharp-sdk/examples/src/SeedExamples/File/Service/Requests/GetFileRequest.cs
@@ -6,6 +6,8 @@ namespace SeedExamples.File;
 
 public record GetFileRequest
 {
+    public required string XFileApiVersion { get; set; }
+
     public override string ToString()
     {
         return JsonUtils.Serialize(this);