+++ seed/csharp-sdk/examples/src/SeedExamples/Types/Types/ResponseType.cs
@@ -9,7 +9,6 @@ namespace SeedExamples;
 public record ResponseType
 {
     [JsonPropertyName("type")]
-    [JsonConverter(typeof(OneOfSerializer<OneOf<BasicType, ComplexType>>))]
     public required OneOf<BasicType, ComplexType> Type { get; set; }
 
     public override string ToString()