+++ seed/csharp-sdk/exhaustive/no-root-namespace-for-core-classes/src/SeedExhaustive/ReqWithHeaders/Requests/ReqWithHeaders.cs
@@ -8,6 +8,8 @@ public record ReqWithHeaders
 {
     public required string XTestEndpointHeader { get; set; }
 
+    public required string XTestServiceHeader { get; set; }
+
     public required string Body { get; set; }
 
     public override string ToString()